misc/HtmlTestSuite/results/test-0140.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Inline Layout: parsing vertical-align with valid values</title>
<link href="https://www.w3.org/TR/CSS2/visudet.html#propdef-vertical-align" rel="help"/>
<link href="https://drafts.csswg.org/css-inline/#propdef-vertical-align" rel="help"/>
<meta content="vertical-align supports the CSS 2 grammar 'baseline | sub | super | top | text-top | middle | bottom | text-bottom | &lt;percentage&gt; | &lt;length&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("vertical-align", "baseline");
test_valid_value("vertical-align", "sub");
test_valid_value("vertical-align", "super");
test_valid_value("vertical-align", "top");
test_valid_value("vertical-align", "text-top");
test_valid_value("vertical-align", "middle");
test_valid_value("vertical-align", "bottom");
test_valid_value("vertical-align", "text-bottom");
test_valid_value("vertical-align", "-10px");
test_valid_value("vertical-align", "20%");
test_valid_value("vertical-align", "0", "0px");
test_valid_value("vertical-align", "calc(20% - 10px)");

// TODO: Test against the CSS Inline Layout grammar when browsers support it.
</script>
</body>
</html>
